Updated package path for matcher extension file.

+++ src/main/java/won/bot/debugbot/impl/DebugBot.java
@@ -26,7 +26,6 @@ import won.bot.framework.eventbot.bus.EventBus;
 import won.bot.framework.eventbot.event.Event;
 import won.bot.framework.eventbot.event.impl.command.close.CloseCommandSuccessEvent;
 import won.bot.framework.eventbot.event.impl.lifecycle.ActEvent;
-import won.bot.framework.eventbot.event.impl.matcher.MatcherExtensionAtomCreatedEvent;
 import won.bot.framework.eventbot.event.impl.wonmessage.*;
 import won.bot.framework.eventbot.filter.impl.AtomUriInNamedListFilter;
 import won.bot.framework.eventbot.filter.impl.NotFilter;
@@ -35,6 +34,7 @@ import won.bot.framework.eventbot.listener.EventListener;
 import won.bot.framework.eventbot.listener.impl.ActionOnEventListener;
 import won.bot.framework.extensions.matcher.MatcherBehaviour;
 import won.bot.framework.extensions.matcher.MatcherExtension;
+import won.bot.framework.extensions.matcher.MatcherExtensionAtomCreatedEvent;
 import won.protocol.model.SocketType;
 
 import java.net.URI;
@@ -54,7 +54,7 @@ public class DebugBot extends EventBot implements MatcherExtension {
     protected BaseEventListener autoOpener;
     protected BaseEventListener messageFromOtherAtomListener;
     protected BaseEventListener usageMessageSender;
-    private int registrationMatcherRetryInterval = 30000;
+    private long registrationMatcherRetryInterval = 30000L;
     private MatcherBehaviour matcherBehaviour;
 
     public void setRegistrationMatcherRetryInterval(final int registrationMatcherRetryInterval) {
